% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinytest.R
\name{expect_snapshot}
\alias{expect_snapshot}
\title{\code{tinytest} helper}
\usage{
expect_snapshot(label, current)
}
\arguments{
\item{label}{Snapshot name}

\item{current}{Code that outputs some text}
}
\description{
\code{tinytest} helper
}
\keyword{internal}
