% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node-traversal}
\alias{node-traversal}
\alias{node_parent}
\alias{node_child}
\alias{node_field}
\alias{node_ancestors}
\alias{node_children}
\alias{node_next}
\alias{node_next_all}
\alias{node_prev}
\alias{node_prev_all}
\title{Navigate the tree}
\usage{
node_parent(x)

node_child(x, nth)

node_field(name)

node_ancestors(x)

node_children(x)

node_next(x)

node_next_all(x)

node_prev(x)

node_prev_all(x)
}
\arguments{
\item{x}{A node, either from \code{\link[=tree_root]{tree_root()}} or from another \verb{node_*()}
function.}

\item{nth}{Integer. The child node to find. This is 0-indexed, so setting
\code{nth = 0} gets the first child.}
}
\description{
This is a collection of functions used to navigate the tree. Some of
them have a variant that applies on a single node (e.g. \code{node_next()}) and
one that applies on a list of nodes (e.g. \code{node_next_all()}).
}
