% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree_new}
\alias{tree_new}
\title{Create a syntax tree}
\usage{
tree_new(txt, file)
}
\arguments{
\item{txt}{A character string of length 1 containing the code to parse.
If provided, \code{file} must not be provided.}

\item{file}{Path to file containing the code to parse. If provided, \code{txt}
must not be provided.}
}
\description{
This function takes R code as string and creates the corresponding abstract
syntax tree (AST) from which we can query nodes.
}
\examples{
src <- "x <- rnorm(100, mean = 2)
    any(duplicated(y))
    plot(x)
    any(duplicated(x))"

tree_new(src)
}
