% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree_root}
\alias{tree_root}
\title{Get the root of the syntax tree}
\usage{
tree_root(x)
}
\arguments{
\item{x}{A tree created by \code{\link[=tree_new]{tree_new()}}.}
}
\value{
A node corresponding to the root of the abstract syntax tree
}
\description{
This function takes a tree created by \code{\link[=tree_new]{tree_new()}} and returns the root node
containing all subsequent nodes.
}
\examples{
src <- "x <- rnorm(100, mean = 2)
    any(duplicated(y))
    plot(x)
    any(duplicated(x))"

tree <- tree_new(src)
tree_root(tree)
}
