% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{astgrepr-package}
\alias{astgrepr}
\alias{astgrepr-package}
\title{astgrepr: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etiennebacher/astgrepr}
  \item Report bugs at \url{https://github.com/etiennebacher/astgrepr/issues}
}

}
