% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{astgrepr-package}
\alias{astgrepr}
\alias{astgrepr-package}
\title{astgrepr: Parse and Manipulate R Code}
\description{
Parsing R code is key to build tools such as linters and stylers. This package provides a binding to the Rust crate 'ast-grep' so that one can parse and explore R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/etiennebacher/astgrepr}
  \item Report bugs at \url{https://github.com/etiennebacher/astgrepr/issues}
}

}
\author{
\strong{Maintainer}: Etienne Bacher \email{etienne.bacher@protonmail.com} [copyright holder]

}
