% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node-find}
\alias{node-find}
\alias{node_find}
\alias{node_find_all}
\title{Find node(s) matching a pattern}
\usage{
node_find(
  x,
  pattern = NULL,
  kind = NULL,
  regex = NULL,
  inside = NULL,
  has = NULL,
  precedes = NULL,
  follows = NULL,
  all = NULL,
  any = NULL,
  not = NULL,
  matches = NULL
)

node_find_all(x, pattern, config = NULL)
}
\arguments{
\item{x}{A node, either from \code{\link[=tree_root]{tree_root()}} or from another \verb{node_*()}
function.}

\item{pattern}{The pattern to search. This must be a list of named elements.
Elements can be:
\itemize{
\item \code{pattern}:
\item \code{kind}:
}}
}
\value{
\code{node_find()} returns a single \code{SgNode}.

\code{node_find_all()} returns a list of \code{SgNode}s.
}
\description{
Those functions find one or several nodes based on some rule:
\itemize{
\item \code{node_find()} returns the first node that is found;
\item \code{node_find_all()} returns a list of all nodes found.
}
}
\examples{
src <- "x <- rnorm(100, mean = 2)
    any(duplicated(y))
    plot(mtcars)
    any(duplicated(x))"

root <- src |>
  tree_new() |>
  tree_root()

root |>
  node_find(pattern = "any(duplicated($A))")

root |>
  node_find_all(list(
    pattern = "any(duplicated($A))"
  ))
}
