% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node-fix}
\alias{node-fix}
\alias{node_replace}
\alias{node_replace_all}
\alias{node_commit_edits}
\title{Change the code in the tree}
\usage{
node_replace(x, ...)

node_replace_all(x, new_text)

node_commit_edits(x, edits)
}
\arguments{
\item{x}{A node, either from \code{\link[=tree_root]{tree_root()}} or from another \verb{node_*()}
function.}

\item{new_text}{The replacement for a node.}

\item{edits}{A list of replacements (the output of \code{node_replace()} or
\code{node_replace_all()}).}
}
\description{
\code{node_replace()} stores the replacement for a particular node, it doesn't
change the actual code. \code{node_replace_all()} does the same but for several
nodes (e.g. the output of \code{node_find_all()}).

\code{node_commit_edits()} takes a list of replacements (the output of
\code{node_replace()} or \code{node_replace_all()}) and applies them one by one to the
input node, returning the modified code.
}
\examples{
src <- "x <- rnorm(100, mean = 2)
any(duplicated(y))
plot(mtcars)
any(duplicated(x))"

root <- src |>
  tree_new() |>
  tree_root()

# one replacement ------------------------------------------

to_fix <- root |>
  node_find(ast_rule(pattern = "any(duplicated($A))"))

fix <- to_fix |>
  node_replace(
    paste0(
      "anyDuplicated(",
      node_text(node_get_match(to_fix, "A")),
      ") > 0"
    )
  )

fixed <- node_commit_edits(root, fix)
cat(fixed[[1]])
}
