% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node_kind}
\alias{node_kind}
\title{Find the kind of a node}
\usage{
node_kind(x)
}
\arguments{
\item{x}{A node, either from \code{\link[=tree_root]{tree_root()}} or from another \verb{node_*()}
function.}
}
\description{
Find the kind of a node
}
\examples{
src <- "x <- rnorm(100, mean = 2)
    any(duplicated(y))
    x <- z + 1
    any(duplicated(x))"

root <- src |>
  tree_new() |>
  tree_root()

root |>
  node_find(pattern = "any(duplicated($VAR))") |>
  node_kind()

root |>
  node_find(pattern = "$X + $VALUE") |>
  node_kind()
}
