% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\name{node_kind}
\alias{node_kind}
\title{Find the kind of a node}
\usage{
node_kind(x)
}
\arguments{
\item{x}{A node, either from \code{\link[=tree_root]{tree_root()}} or from another \verb{node_*()}
function.}
}
\description{
Find the kind of a node
}
