% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\name{relational_rule}
\alias{relational_rule}
\title{Build a relational rule}
\usage{
relational_rule(stopBy = "neighbor", field = NULL, regex = NULL)
}
\arguments{
\item{stopBy}{todo}

\item{field}{todo}

\item{regex}{todo}
}
\description{
Build a relational rule
}
