% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tree_new}
\alias{tree_new}
\title{Create a syntax tree}
\usage{
tree_new(src)
}
\arguments{
\item{src}{String. R code to parse.}
}
\description{
This function takes R code as string and creates the corresponding abstract
syntax tree (AST) from which we can query nodes.
}
\examples{
src <- "x <- rnorm(100, mean = 2)
    any(duplicated(y))
    plot(x)
    any(duplicated(x))"

tree_new(src)
}
